:recycle: Remove package classes from DetailProductResponse (but they are still package classes)
