+++ src/main/java/com/mercadolivre/apimlv2/usecases/detailsproduct/DetailProductResponse.java
@@ -1,6 +1,8 @@
 package com.mercadolivre.apimlv2.usecases.detailsproduct;
 
-import com.mercadolivre.apimlv2.domain.*;
+import com.mercadolivre.apimlv2.domain.Opinion;
+import com.mercadolivre.apimlv2.domain.Product;
+import com.mercadolivre.apimlv2.domain.ProductImage;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -27,59 +29,22 @@ public class DetailProductResponse {
         this.name = product.getName();
         this.price = product.getPrice();
         this.amountAvailable = product.getAmountAvailable();
-        this.features = product.getFeatures().stream().map(FeatureResponse::new).collect(Collectors.toList());
+        this.features = product.getFeatures().stream()
+                               .map(FeatureResponse::new)
+                               .collect(Collectors.toList());
         this.description = product.getDescription();
         this.category = new CategoryResponse(product.getCategory());
         this.owner = new OwnerResponse(product.getOwner());
         this.createdAt = product.getCreatedAt();
-        this.images = product.getImages().stream().map(ProductImage::getImageUrl).collect(Collectors.toList());
-        this.opinions = product.getOpinions().stream().map(OpinionResponse::new).collect(Collectors.toList());
+        this.images = product.getImages().stream()
+                             .map(ProductImage::getImageUrl)
+                             .collect(Collectors.toList());
+        this.opinions = product.getOpinions().stream()
+                               .map(OpinionResponse::new)
+                               .collect(Collectors.toList());
         this.amountOfOpinions = product.getOpinions().size();
-        this.scoreAverage = product.getOpinions().stream().mapToInt(Opinion::getScore).average().orElse(0);
-    }
-
-}
-
-class FeatureResponse {
-
-    public final String name;
-    public final String description;
-
-    public FeatureResponse(ProductFeature feature) {
-        this.name = feature.getName();
-        this.description = feature.getDescription();
-    }
-}
-
-class CategoryResponse {
-
-    public final String name;
-
-    public CategoryResponse(Category category) {
-        this.name = category.getName();
-    }
-}
-
-class OwnerResponse {
-
-    public final String email;
-
-    public OwnerResponse(User owner) {
-        this.email = owner.getLogin();
-    }
-}
-
-class OpinionResponse {
-
-    public final String title;
-    public final String description;
-    public final String opinionatorEmail;
-    public final int score;
-
-    public OpinionResponse(Opinion opinion) {
-        title = opinion.getTitle();
-        description = opinion.getDescription();
-        opinionatorEmail = opinion.getOpinionator().getLogin();
-        score = opinion.getScore();
+        this.scoreAverage = product.getOpinions().stream()
+                                   .mapToInt(Opinion::getScore)
+                                   .average().orElse(0);
     }
 }
